api: rename proxy to ollama
